refactor : provider를 통한 코드 중복 제거 (#37)
