+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/domain/bakeryfavorite/controller/req/GeoPointRequest.java
@@ -1,6 +1,6 @@
 package com.breaditnow.customer.domain.bakeryfavorite.controller.req;
 
-import com.breaditnow.common.util.geodistance.GeoPoint;
+import com.breaditnow.common.util.GeoPoint;
 
 import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.DecimalMin;
