+++ breaditnow-domain/src/main/java/com/breaditnow/domain/domain/favorite/repository/customerbakeryfavorite/CustomerBakeryFavoriteRepositoryCustom.java
@@ -3,7 +3,7 @@ package com.breaditnow.domain.domain.favorite.repository.customerbakeryfavorite;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import com.breaditnow.common.util.geodistance.GeoPoint;
+import com.breaditnow.common.util.GeoPoint;
 import com.breaditnow.domain.domain.favorite.dto.BakeryFavoriteDto;
 
 public interface CustomerBakeryFavoriteRepositoryCustom {
