+++ breaditnow-domain/src/main/java/com/breaditnow/domain/domain/favorite/repository/customerbakeryfavorite/CustomerBakeryFavoriteRepositoryImpl.java
@@ -3,7 +3,6 @@ package com.breaditnow.domain.domain.favorite.repository.customerbakeryfavorite;
 import static com.breaditnow.domain.domain.bakery.entity.QBakery.*;
 import static com.breaditnow.domain.domain.favorite.entity.QCustomerBakeryFavorite.*;
 import static com.querydsl.core.types.Projections.*;
-import static com.querydsl.core.types.dsl.Expressions.*;
 
 import java.util.List;
 
@@ -11,12 +10,12 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 
-import com.breaditnow.common.util.geodistance.GeoPoint;
+import com.breaditnow.common.util.GeoPoint;
 import com.breaditnow.domain.domain.favorite.dto.BakeryFavoriteDto;
+import com.breaditnow.domain.domain.favorite.repository.GeoDistanceExpressionProvider;
 import com.breaditnow.domain.domain.favorite.repository.customerbakeryfavorite.strategy.BakeryFavoriteSortStrategy;
 import com.breaditnow.domain.domain.favorite.repository.customerbakeryfavorite.strategy.BakeryFavoriteSortStrategyFactory;
 import com.querydsl.core.types.dsl.BooleanExpression;
-import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.core.types.dsl.NumberExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
@@ -26,6 +25,7 @@ import lombok.RequiredArgsConstructor;
 public class CustomerBakeryFavoriteRepositoryImpl implements CustomerBakeryFavoriteRepositoryCustom {
 	private final BakeryFavoriteSortStrategyFactory sortFactory;
 	private final JPAQueryFactory queryFactory;
+	private final GeoDistanceExpressionProvider distanceExpressionProvider;
 
 	@Override
 	public Page<BakeryFavoriteDto> findBakeryFavorites(Long customerId, Pageable pageable, GeoPoint geoPoint) {
@@ -34,10 +34,12 @@ public class CustomerBakeryFavoriteRepositoryImpl implements CustomerBakeryFavor
 			.and(customerBakeryFavorite.isActive.eq(true));
 
 		BakeryFavoriteSortStrategy strategy = sortFactory.getStrategy(pageable.getSort(), geoPoint);
+		NumberExpression<Double> distanceExpression = distanceExpressionProvider.buildDistanceExpression(geoPoint,
+			bakery);
 
 		List<BakeryFavoriteDto> content = queryFactory
 			.select(
-				constructor(BakeryFavoriteDto.class, bakery, getDist(geoPoint))
+				constructor(BakeryFavoriteDto.class, bakery, distanceExpression)
 			)
 			.from(customerBakeryFavorite, bakery)
 			.where(baseCondition)
@@ -55,20 +57,5 @@ public class CustomerBakeryFavoriteRepositoryImpl implements CustomerBakeryFavor
 			.fetchOne();
 
 		return new PageImpl<>(content, pageable, totalCount == null ? 0 : totalCount);
-	}
-
-	private NumberExpression<Double> getDist(GeoPoint geoPoint) {
-		if (geoPoint == null) {
-			return Expressions.asNumber(Expressions.nullExpression(Double.class));
-		}
-
-		return Expressions.numberTemplate(
-			Double.class,
-			"cast(function('ST_Distance_Sphere', function('Point', {0}, {1}), function('Point', {2}, {3})) as double) / 1000",
-			constant(geoPoint.longitude()),
-			constant(geoPoint.latitude()),
-			bakery.address.longitude,
-			bakery.address.latitude
-		);
 	}
 }
