+++ breaditnow-domain/src/main/java/com/breaditnow/domain/domain/favorite/repository/customerbakeryfavorite/strategy/BakeryFavoriteSortStrategy.java
@@ -1,6 +1,6 @@
 package com.breaditnow.domain.domain.favorite.repository.customerbakeryfavorite.strategy;
 
-import com.breaditnow.common.util.geodistance.GeoPoint;
+import com.breaditnow.common.util.GeoPoint;
 import com.breaditnow.domain.domain.bakery.entity.QBakery;
 import com.breaditnow.domain.domain.favorite.entity.QCustomerBakeryFavorite;
 import com.querydsl.core.types.OrderSpecifier;
