+++ breaditnow-domain/src/main/java/com/breaditnow/domain/domain/favorite/repository/customerbakeryfavorite/strategy/DistanceBakeryFavoriteSortStrategy.java
@@ -1,33 +1,29 @@
 package com.breaditnow.domain.domain.favorite.repository.customerbakeryfavorite.strategy;
 
 import static com.breaditnow.domain.global.exception.DomainErrorCode.*;
-import static com.querydsl.core.types.dsl.Expressions.*;
 
 import org.springframework.stereotype.Component;
 
-import com.breaditnow.common.util.geodistance.GeoPoint;
+import com.breaditnow.common.util.GeoPoint;
 import com.breaditnow.domain.domain.bakery.entity.QBakery;
 import com.breaditnow.domain.domain.favorite.entity.QCustomerBakeryFavorite;
+import com.breaditnow.domain.domain.favorite.repository.GeoDistanceExpressionProvider;
 import com.breaditnow.domain.global.exception.DomainException;
 import com.querydsl.core.types.OrderSpecifier;
-import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.core.types.dsl.NumberExpression;
 
+import lombok.RequiredArgsConstructor;
+
 @Component
+@RequiredArgsConstructor
 public class DistanceBakeryFavoriteSortStrategy implements BakeryFavoriteSortStrategy {
 	private GeoPoint currentGeoPoint;
+	private final GeoDistanceExpressionProvider distanceExpressionProvider;
 
 	@Override
 	public OrderSpecifier<?> getOrderSpecifier(QBakery bakery, QCustomerBakeryFavorite customerBakeryFavorite) {
-		NumberExpression<Double> distanceExpression = Expressions.numberTemplate(
-			Double.class,
-			"cast(function('ST_Distance_Sphere', function('Point', {0}, {1}), function('Point', {2}, {3})) as double) / 1000",
-			constant(currentGeoPoint.longitude()),
-			constant(currentGeoPoint.latitude()),
-			bakery.address.longitude,
-			bakery.address.latitude
-		);
-
+		NumberExpression<Double> distanceExpression = distanceExpressionProvider.buildDistanceExpression(
+			currentGeoPoint, bakery);
 		return distanceExpression.asc();
 	}
